refactor(dashboard): Move chart zoom into a composable
BarChart carried the drag-to-zoom setup inline. LineChart needs the same
behaviour, so the setup moves to useDataZoom, which the chart gives a
function to turn an axis value into a date.

The event shape is now read as `event.batch?.[0] ?? event`. A category
axis reports the picked range in a batch, a time axis reports it on the
event itself.
